requirements.txt
wheel
pytest
sqlalchemy
psycopg2
flask
marshmallow
python-abc
requests
pytest-cov
pytest-mock
pytest-env
flask_wtf
argon2-cffi
flask-jwt-extended